+++ idea/src/org/jetbrains/kotlin/idea/quickfix/createFromUsage/createTypeAlias/CreateTypeAliasFromTypeReferenceActionFactory.kt
@@ -16,30 +16,18 @@
 
 package org.jetbrains.kotlin.idea.quickfix.createFromUsage.createTypeAlias
 
-import org.jetbrains.kotlin.descriptors.TypeParameterDescriptor
 import org.jetbrains.kotlin.diagnostics.Diagnostic
-import org.jetbrains.kotlin.idea.caches.resolve.analyze
 import org.jetbrains.kotlin.idea.core.CollectingNameValidator
 import org.jetbrains.kotlin.idea.core.KotlinNameSuggester
 import org.jetbrains.kotlin.idea.core.NewDeclarationNameValidator
 import org.jetbrains.kotlin.idea.quickfix.createFromUsage.callableBuilder.contains
 import org.jetbrains.kotlin.idea.quickfix.createFromUsage.createClass.CreateClassFromTypeReferenceActionFactory
-import org.jetbrains.kotlin.psi.*
+import org.jetbrains.kotlin.idea.quickfix.createFromUsage.createClass.getBoundingTypeParameter
+import org.jetbrains.kotlin.psi.KtUserType
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfTypeAndBranch
-import org.jetbrains.kotlin.resolve.bindingContextUtil.getAbbreviatedTypeOrType
-import org.jetbrains.kotlin.resolve.lazy.BodyResolveMode
 
 object CreateTypeAliasFromTypeReferenceActionFactory : CreateTypeAliasFromUsageFactory<KtUserType>() {
     override fun getElementOfInterest(diagnostic: Diagnostic) = CreateClassFromTypeReferenceActionFactory.getElementOfInterest(diagnostic)
-
-    private fun getBoundingTypeParameter(element: KtUserType): TypeParameterDescriptor? {
-        val context = element.analyze(BodyResolveMode.PARTIAL)
-        val containingTypeArg = (element.parent as? KtTypeReference)?.parent as? KtTypeProjection ?: return null
-        val argumentList = containingTypeArg.parent as? KtTypeArgumentList ?: return null
-        val containingTypeRef = (argumentList.parent as? KtTypeElement)?.parent as? KtTypeReference ?: return null
-        val containingType = containingTypeRef.getAbbreviatedTypeOrType(context) ?: return null
-        return containingType.constructor.parameters.getOrNull(argumentList.arguments.indexOf(containingTypeArg))
-    }
 
     override fun extractFixData(element: KtUserType, diagnostic: Diagnostic): TypeAliasInfo? {
         if (element.getParentOfTypeAndBranch<KtUserType>(true) { qualifier } != null) return null