+++ idea/testData/quickfix/createFromUsage/createTypeParameter/functionNoExplication.kt
@@ -1,0 +1,9 @@
+// "Create type parameter 'X'" "true"
+fun foo(x: <caret>X) {
+
+}
+
+fun test() {
+    foo(1)
+    foo("2")
+}