+++ idea/testData/quickfix/createFromUsage/createTypeParameter/qualifiedType.kt
@@ -1,0 +1,13 @@
+// "Create type parameter 'X'" "false"
+// ACTION: Create annotation 'X'
+// ACTION: Create class 'X'
+// ACTION: Create enum 'X'
+// ACTION: Create interface 'X'
+// ACTION: Create type alias 'X'
+// ERROR: Unresolved reference: X
+
+class A
+
+fun foo(x: A.<caret>X) {
+
+}