+++ idea/testData/quickfix/typeMismatch/typeMismatchOnReturnedExpression/dontChangeFunctionReturnTypeToErrorType.kt
@@ -5,6 +5,7 @@
 // ACTION: Create interface 'NoSuchType'
 // ACTION: Create type alias 'NoSuchType'
 // ACTION: Remove explicit lambda parameter types (may break code)
+// ACTION: Create type parameter 'NoSuchType'
 // ERROR: Type mismatch: inferred type is ([ERROR : NoSuchType]) -> Int but Int was expected
 // ERROR: Unresolved reference: NoSuchType